+++ src/Contracts/RepositoryInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Prozorov\Repositories\Contracts;
+namespace Repositories\Core\Contracts;
 
 interface RepositoryInterface
 {
@@ -28,28 +28,28 @@ interface RepositoryInterface
      * Get model by id
      *
      * @access	public
-     * @param	int  	$id
+     * @param	mixed  	$id
      * @param	array	$select	Default: null
      * @return	mixed
      */
-    public function getById(int $id, array $select = null);
+    public function getById($id, array $select = null);
 
     /**
      * Get model or fail
      *
      * @access	public
-     * @param	int  	$id    	
+     * @param	mixed  	$id    	
      * @param	array	$select	Default: null
      * @return	mixed
      */
-    public function getByIdOrFail(int $id, array $select = null);
+    public function getByIdOrFail($id, array $select = null);
 
     /**
      * Create a new record
      *
      * @param array $data
      * @param array|null $guarded
-     * @throws \Prozorov\Repositories\Exceptions\RepositoryException
+     * @throws \Repositories\Core\Exceptions\RepositoryException
      * @return mixed
      */
     public function create(array $data);
@@ -59,7 +59,7 @@ interface RepositoryInterface
      *
      * @param $model
      * @param array $data
-     * @throws \Prozorov\Repositories\Exceptions\RepositoryException
+     * @throws \Repositories\Core\Exceptions\RepositoryException
      * @return void
      */
     public function update($model, array $data): void;
@@ -69,7 +69,7 @@ interface RepositoryInterface
      *
      * @access	public
      * @param	mixed	$model	
-     * @throws \Prozorov\Repositories\Exceptions\RepositoryException
+     * @throws \Repositories\Core\Exceptions\RepositoryException
      * @return	void
      */
     public function delete($model): void;
@@ -95,7 +95,7 @@ interface RepositoryInterface
      * Open transaction
      *
      * @access	public
-     * @throws \Prozorov\Repositories\Exceptions\RepositoryException
+     * @throws \Repositories\Core\Exceptions\RepositoryException
      * @return	void
      */
     public function openTransaction(): void;
@@ -104,7 +104,7 @@ interface RepositoryInterface
      * Commit transaction if one exists
      *
      * @access	public
-     * @throws \Prozorov\Repositories\Exceptions\RepositoryException
+     * @throws \Repositories\Core\Exceptions\RepositoryException
      * @return	void
      */
     public function commitTransaction(): void;
@@ -113,8 +113,29 @@ interface RepositoryInterface
      * Rollback transaction if one exists
      *
      * @access	public
-     * @throws \Prozorov\Repositories\Exceptions\RepositoryException
+     * @throws \Repositories\Core\Exceptions\RepositoryException
      * @return	void
      */
     public function rollbackTransaction(): void;
+
+    /**
+     * Insert multiple entries into the database. No transaction is active.
+     *
+     * @access	public
+     * @param	iterable $data
+     * @throws \Repositories\Core\Exceptions\RepositoryException
+     * @return	void
+     */
+    public function insert(iterable $data): void;
+
+    /**
+     * Insert multiple entries into the database.
+     * This operation is transaction-controlled so in case of any error all data will be rolled back.
+     *
+     * @access	public
+     * @param	iterable $data
+     * @throws \Repositories\Core\Exceptions\RepositoryException
+     * @return	void
+     */
+    public function insertWithTransaction(iterable $data): void;
 }