+++ src/Exceptions/RepositoryException.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Prozorov\Repositories\Exceptions;
+namespace Repositories\Core\Exceptions;
 
 use InvalidArgumentException;