+++ tests/Resolvers/ChainResolverTest.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Prozorov\Repositories\Tests\Resolvers;
+namespace Repositories\Core\Tests\Resolvers;
 
 use Mockery\Adapter\Phpunit\MockeryTestCase;
-use Prozorov\Repositories\Resolvers\{ChainResolver, ContainerAwareResolver, HardResolver};
-use Prozorov\Repositories\Contracts\{RepositoryInterface, ResolverInterface};
-use Prozorov\Repositories\Exceptions\CouldNotResolve;
-use Prozorov\Repositories\Tests\Support\UserRepository;
+use Repositories\Core\Resolvers\{ChainResolver, ContainerAwareResolver, HardResolver};
+use Repositories\Core\Contracts\{RepositoryInterface, ResolverInterface};
+use Repositories\Core\Exceptions\CouldNotResolve;
+use Repositories\Core\Tests\Support\UserRepository;
 
 class ChainResolverTest extends MockeryTestCase
 {